[#168761853] Pass emailService as a parameter to newApp()
